Move shared test infrastructure into tests/common.
