+++ tests/unit/test_build_schemas/test_setuptools_build_schema.py
@@ -1,7 +1,7 @@
 from quickpub import SetuptoolsBuildSchema
 
-from tests.base_test_classes import BaseTestClass
-from tests.test_helpers import temporary_test_directory
+from tests.common.base_test_classes import BaseTestClass
+from tests.common.helpers import temporary_test_directory
 
 TMP_SETUP_FILE_PATH: str = "tmp_setup.py"
 TMP_TOML_FILE_PATH: str = "tmp_pyproject.toml"
