+++ tests/unit/test_constraint_enforcers/test_pypirc_enforcer.py
@@ -1,7 +1,7 @@
 from quickpub import PypircEnforcer
 
-from tests.base_test_classes import BaseTestClass
-from tests.test_helpers import temporary_test_directory
+from tests.common.base_test_classes import BaseTestClass
+from tests.common.helpers import temporary_test_directory
 
 TMP_PYPIRC_PATH: str = ".TMP_PYPIRC"
