+++ tests/unit/test_files.py
@@ -20,8 +20,8 @@ from quickpub.files import (
     _build_toml_content,
 )
 
-from tests.base_test_classes import BaseTestClass
-from tests.test_helpers import temporary_test_directory
+from tests.common.base_test_classes import BaseTestClass
+from tests.common.helpers import temporary_test_directory
 
 
 class TestFormatClassifiersString(BaseTestClass):
