+++ tests/unit/test_python_providers/test_default_python_provider.py
@@ -4,8 +4,8 @@ from danielutils import get_python_version
 
 from quickpub import DefaultPythonProvider
 
-from tests.base_test_classes import AsyncBaseTestClass
-from tests.test_helpers import temporary_test_directory
+from tests.common.base_test_classes import AsyncBaseTestClass
+from tests.common.helpers import temporary_test_directory
 
 
 class TestDefaultPythonProvider(AsyncBaseTestClass):
