+++ tests/unit/test_qa_runners/test_pytest_runner.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock, patch
 
 from quickpub import DefaultPythonProvider, PytestRunner, ExitEarlyError
 
-from tests.base_test_classes import AsyncBaseTestClass, BaseTestClass
-from tests.test_helpers import temporary_test_directory
+from tests.common.base_test_classes import AsyncBaseTestClass, BaseTestClass
+from tests.common.helpers import temporary_test_directory
 
 TEST_FILE_PATH: str = "test_foo.py"
